+++ DreamControlApp/Onboarding/MainOnboarding.swift
@@ -1,0 +1,25 @@
+//
+//  MainOnboarding.swift
+//  DreamControlApp
+//
+//  Created by Ainur on 05.03.2024.
+//
+
+import SwiftUI
+
+struct MainOnboarding: View {
+    var body: some View {
+        TabView {
+            OnboardingView()
+            OnboardingView2()
+            OnboardingView3()
+
+        }
+        .tabViewStyle(.page)
+        .indexViewStyle(.page(backgroundDisplayMode: .always))
+    }
+}
+
+#Preview {
+    MainOnboarding()
+}
