+++ DreamControlApp/SplashView.swift
@@ -15,7 +15,7 @@ struct SplashView: View {
     
     var body: some View {
         if isActive {
-            OnboardingView()
+            MainOnboarding()
         } else {
             VStack {
                 VStack {
